BulkLoadVisitor: use the original request data when processing results
When the data source is a database query, there's no guarantee that we
will get the same result set back when we run the same query a second
time. Instead of re-reading the data source, query the server to
retrieve the CSV data the was sent in the original batch request.